Correció de calcul de la mitjana.
Correció de calcul de la mitjana.

+++ Exercici 7/functions.h
@@ -6,10 +6,10 @@
 
 float validar();
 float validar2();
-void omplir(float *, float);
-float mitjana(float *, float);
-float max(float *, float);
-float min(float *, float);
+void omplir(float *, int);
+float mitjana(float *, int);
+float max(float *, int);
+float min(float *, int);
 
 
 #endif /* FUNCTIONS_H_ */
